Added expandIn operator, fixed relation conflicts
Added expandIn operator, that takes a relation and a target node and
expands the relation to include all edges that end in target.
Fixed how I handled relations: previously i was updating the rows
themselves and so when I had rows with duplicate keys (like 2 al
pacinos) they would overwrite eachother.
Now I build a new relationship every time and each row is a new dict.
--- TO CHECK: this keeps only nodes that were relevant to our operator -
i.e i may lose rows that don't conform to the new schema (for example
when i expandOut, if some row doesn't contain the source node do i lose
it? If it remains, should I add NULL to the fields that it is missing?